Implémentation des enumerations sur les champs Long

+++ mts/src/main/java/com/devoteam/srit/xmlloader/gtp/data/HeaderGTPV1.java
@@ -116,10 +116,10 @@ public class HeaderGTPV1 extends HeaderAbstract
         if (strType != null)
         {
             EnumerationField field = (EnumerationField) dictionary.getHeaderFieldByName("Message Type");
-            this.type = field.getEnumValue(strType);
+            this.type = (int) field.getEnumValue(strType);
         }
         EnumerationField field = (EnumerationField) dictionary.getHeaderFieldByName("Message Type");
-        this.label = field.getEnumLabelByValue(this.type);
+        this.label = field.getEnumLabelByValue((long) this.type);
         
         String attribute;
         String attrFlag;
@@ -279,7 +279,7 @@ public class HeaderGTPV1 extends HeaderAbstract
 		int offset = 4;
 			
     	EnumerationField field = (EnumerationField) dictionary.getHeaderFieldByName("Message Type");
-	    this.label = field.getEnumLabelByValue(this.type);    	
+	    this.label = field.getEnumLabelByValue((long) this.type);    	
 
 	    Array teidArray = array.subArray(offset, 4); 
         this.tunnelEndpointId = new Integer32Array(teidArray).getValue() & 0xffffffffl;
